pyproject.toml
[project]
name = "document-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-ai-contentunderstanding>=1.0.1",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.38.2",
    "azure-identity>=1.25.2",
    "azure-storage-blob>=12.28.0",
    "python-dotenv>=1.2.2",
]
